+++ src/crawl4md/__init__.py
@@ -1,6 +1,8 @@
 from .config import MarkdownPreprocessingConfig, ParseType
 from .convert.crawl4ai_markdown import Crawl4AIMarkdownConverter
+from .convert.kreuzberg_dev_markdown import KreuzbergDevMarkdownConverter
 from .fetch.crawl4ai_markdown import Crawl4AIMarkdownFetcher
+from .fetch.kreuzberg_dev_markdown import KreuzbergDevMarkdownFetcher
 
 
 MarkdownConverter = Crawl4AIMarkdownConverter
@@ -10,6 +12,8 @@ MarkdownFetcher = Crawl4AIMarkdownFetcher
 __all__ = [
     "Crawl4AIMarkdownConverter",
     "Crawl4AIMarkdownFetcher",
+    "KreuzbergDevMarkdownConverter",
+    "KreuzbergDevMarkdownFetcher",
     "MarkdownConverter",
     "MarkdownFetcher",
     "MarkdownPreprocessingConfig",